chore: use MainScreen on MainActivity
